backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
python-dotenv>=1.0.0
httpx>=0.28.0
openai>=1.54.0
anthropic>=0.40.0
google-genai>=1.0.0
sqlmodel>=0.0.22
aiosqlite>=0.20.0
greenlet>=3.0.0